backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.2
debugpy==1.8.13
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
inflection==0.5.1
MarkupSafe==3.0.2
mysqlclient==2.1.1
packaging==24.2
pycparser==2.22
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
Werkzeug==3.1.3
